Move component style files
