+++ src/components/MapQuestion.tsx
@@ -3,9 +3,9 @@ import { jsx } from '@emotion/core'
 import * as React from 'react'
 import { MapNode } from '../components/MapNode'
 import { getFinalNodeTitle } from '../firestore/firestore'
-import * as styles from '../styles/MapQuestion.style'
+import * as styles from './MapQuestion.style'
 import { knockout, selected, dropShadow } from '../styles/shared.style'
-import { mapNodeChildren } from '../styles/Map.style'
+import { mapNodeChildren } from './MapNode.style'
 import { useTheme } from 'emotion-theming'
 import { Theme } from '@emotion/types'
